Start pymaring.
